# Cache dtype for faster variable access