It seems that many aspects try to access the dtype, and this can slow things down when accessing variables.

I've found that this can speed up certain access times of mine by a factor 2. I'll try to recreate it with the file I provided.

... I really wish it was easy to find small files that recreate this issue, I have to admit, the file I'm benchmarking on is about 16GB in size.  In a read only usecase.

Before

```
============= ================
--               filename 
------------- ----------------
    engine      my_file.nc
============= ================
   netcdf4      19.5±0ms      
   h5netcdf     81.2±0ms      
 h5py_direct    13.1±0ms      
============= ================
```

After
```
============= ================
--               filename 
------------- ----------------
    engine      my_file.nc
============= ================
   netcdf4      19.3±0ms      
   h5netcdf     19.9±0ms      
 h5py_direct    13.1±0ms      
============= ================
```

The benchmark looks something like:

```
import xarray as xr
dataset = xr.open_dataset(filename, engine=engine)
variable = dataset.images.variable
for i in range(30):
    variable[i].data
```
<!-- Feel free to remove check-list items which aren't relevant to your changes -->

- [ ] Closes Issue #xxxx
- [ ] Tests added
- [ ] Changes are documented in `CHANGELOG.rst`